==> notebooks/02_audio_features_librosa.ipynb <==
# 02 Librosa Audio Analysis

import pandas as pd
import numpy as np

df = pd.read_parquet("../data/processed/base_tracks.parquet")

df

df.columns

import requests
import tempfile
import librosa
import os
import warnings
from tqdm import tqdm


tqdm.pandas()
warnings.filterwarnings('ignore')

HEADERS = {
    'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36'
}

def extract_audio_features_final(preview_url):
    if pd.isna(preview_url):
        return None
    
    temp_path = None
    try:
        response = requests.get(preview_url, headers=HEADERS, timeout=10)
        if response.status_code != 200:
            return None
            
        with tempfile.NamedTemporaryFile(suffix=".m4a", delete=False) as temp_file:
            temp_file.write(response.content)
            temp_path = temp_file.name
        
        if os.path.getsize(temp_path) == 0:
            return None

        y, sr = librosa.load(temp_path, duration=30)
        
        tempo, _ = librosa.beat.beat_track(y=y, sr=sr)
        if isinstance(tempo, np.ndarray):
            tempo = tempo.item() if tempo.size > 0 else 0

        rms = np.mean(librosa.feature.rms(y=y))
        spec_cent = np.mean(librosa.feature.spectral_centroid(y=y, sr=sr))
        zcr = np.mean(librosa.feature.zero_crossing_rate(y))
        onset_env = librosa.onset.onset_strength(y=y, sr=sr)
        rhythm_strength = np.mean(onset_env)
        
        if os.path.exists(temp_path):
            os.remove(temp_path)
        
        return {
            "bpm": float(tempo),
            "energy": float(rms),
            "brightness": float(spec_cent),
            "noisiness": float(zcr),
            "rhythm_strength": float(rhythm_strength)
        }

    except Exception:
        if temp_path and os.path.exists(temp_path):
            os.remove(temp_path)
        return None

print(f"🚀 Starting {len(df)} Song Analysis...")

audio_data = df['preview_url'].progress_apply(extract_audio_features_final)

df_audio_features = pd.DataFrame(audio_data.tolist())

expected_cols = ["bpm", "energy", "brightness", "noisiness", "rhythm_strength"]
for col in expected_cols:
    if col not in df_audio_features.columns:
        df_audio_features[col] = np.nan

df_complete = pd.concat([df.reset_index(drop=True), df_audio_features], axis=1)

df_clean = df_complete.dropna(subset=['bpm'])

print("-" * 50)
print(f"🎉🎉🎉 Success!")
print(f"✅ Total {len(df)} | Success {len(df_clean)}")
print(f"✅ Success Rate : {len(df_clean)/len(df)}")
print("-" * 50)

audio_feature_cols = ["bpm", "energy", "brightness", "noisiness", "rhythm_strength"]

df_audio_only = df_complete[["apple_track_id", "preview_url"] + audio_feature_cols].copy()

df_audio_only = df_audio_only.dropna(subset=["bpm"])

df_audio_only

df_audio_only.to_parquet("../data/processed/audio_features.parquet", index=False)
print("💾 Saved as audio_features.parquet")
